refactor: dissolve the appearance domain module
The domain role is gone (ADR 0019): a slice's logic lives in its
services role. The appearance slice still carried a legacy domain
module, so callers had to import the same vocabulary from two
places. Moving the pure code into services closes that gap and
leaves importing as the last slice waiting for its own dissolution.
