+++ mpvqc/appearance/models/accent_colors.py
@@ -10,15 +10,14 @@ import inject
 from PySide6.QtCore import QAbstractListModel, QByteArray, QModelIndex, Qt
 from PySide6.QtQml import QmlElement, QmlUncreatable
 
-from mpvqc.appearance.domain import Dark, FollowSystem, Light
-from mpvqc.appearance.services import PaletteCatalogService
+from mpvqc.appearance.services import Dark, FollowSystem, Light, PaletteCatalogService
 
 if TYPE_CHECKING:
     from typing import Any
 
     from PySide6.QtCore import QObject, QPersistentModelIndex
 
-    from mpvqc.appearance.domain import ColorSchemePreference, Palette
+    from mpvqc.appearance.services import ColorSchemePreference, Palette
 
 QML_IMPORT_NAME = "io.github.mpvqc.mpvQC.Python"
 QML_IMPORT_MAJOR_VERSION = 1
