+++ mpvqc/appearance/services/__init__.py
@@ -4,6 +4,25 @@
 
 from .color_scheme import ColorSchemeService as ColorSchemeService
 from .color_scheme import QtStyleHints as QtStyleHints
+from .palette_catalog import Palette as Palette
 from .palette_catalog import PaletteCatalogService as PaletteCatalogService
 from .palette_catalog import PaletteFamily as PaletteFamily
+from .preferences import AccentColor as AccentColor
+from .preferences import AppearancePreference as AppearancePreference
+from .preferences import NoPreference as NoPreference
+from .schemes import COLOR_SCHEME_PREFERENCES as COLOR_SCHEME_PREFERENCES
+from .schemes import ColorScheme as ColorScheme
+from .schemes import ColorSchemePreference as ColorSchemePreference
+from .schemes import Dark as Dark
+from .schemes import FollowSystem as FollowSystem
+from .schemes import Light as Light
+from .schemes import SystemColorScheme as SystemColorScheme
+from .schemes import Unknown as Unknown
+from .schemes import default_color_scheme_preference as default_color_scheme_preference
+from .schemes import format_color_scheme as format_color_scheme
+from .schemes import format_color_scheme_preference as format_color_scheme_preference
+from .schemes import parse_color_scheme as parse_color_scheme
+from .schemes import parse_color_scheme_preference as parse_color_scheme_preference
+from .schemes import parse_color_scheme_preference_or_default as parse_color_scheme_preference_or_default
+from .schemes import resolve_color_scheme as resolve_color_scheme
 from .settings import AppearanceSettingsService as AppearanceSettingsService
