+++ mpvqc/appearance/services/palette_catalog.py
@@ -9,17 +9,36 @@ from typing import assert_never
 
 import inject
 
-from mpvqc.appearance.domain import (
-    AccentColor,
-    AppearancePreference,
-    ColorScheme,
-    Dark,
-    Light,
-    NoPreference,
-    Palette,
-    parse_color_scheme,
-)
 from mpvqc.services import ResourceService
+
+from .preferences import AccentColor, AppearancePreference, NoPreference
+from .schemes import ColorScheme, Dark, Light, parse_color_scheme
+
+
+@dataclass(frozen=True)
+class Palette:
+    accent_color: AccentColor
+    background: str
+    foreground: str
+    hint: str
+    accent: str
+    separator: str
+    error: str
+    error_text: str
+    header_background: str
+    popup_background: str
+    popup_text: str
+    menu_background: str
+    dialog_background: str
+    section_card: str
+    tooltip_background: str
+    tooltip_text: str
+    row_base: str
+    row_base_text: str
+    row_stripe: str
+    row_stripe_text: str
+    row_selected: str
+    row_selected_text: str
 
 
 def _dark_palette(accent_color: AccentColor, colors: dict[str, str]) -> Palette:
