+++ mpvqc/appearance/services/schemes.py
@@ -7,19 +7,6 @@ from typing import assert_never
 
 
 @dataclass(frozen=True)
-class AccentColor:
-    identifier: str
-
-
-@dataclass(frozen=True)
-class NoPreference:
-    pass
-
-
-type AccentColorPreference = NoPreference | AccentColor
-
-
-@dataclass(frozen=True)
 class Light:
     pass
 
@@ -118,45 +105,3 @@ def resolve_color_scheme(
             return preference
         case _:
             assert_never(preference)
-
-
-@dataclass(frozen=True)
-class AppearancePreference:
-    color_scheme_preference: ColorSchemePreference
-    light_accent_color_preference: AccentColorPreference
-    dark_accent_color_preference: AccentColorPreference
-
-    def accent_color_preference_for(self, color_scheme: ColorScheme) -> AccentColorPreference:
-        match color_scheme:
-            case Light():
-                return self.light_accent_color_preference
-            case Dark():
-                return self.dark_accent_color_preference
-            case _:
-                assert_never(color_scheme)
-
-
-@dataclass(frozen=True)
-class Palette:
-    accent_color: AccentColor
-    background: str
-    foreground: str
-    hint: str
-    accent: str
-    separator: str
-    error: str
-    error_text: str
-    header_background: str
-    popup_background: str
-    popup_text: str
-    menu_background: str
-    dialog_background: str
-    section_card: str
-    tooltip_background: str
-    tooltip_text: str
-    row_base: str
-    row_base_text: str
-    row_stripe: str
-    row_stripe_text: str
-    row_selected: str
-    row_selected_text: str
