+++ pyproject.toml
@@ -125,13 +125,14 @@ files = [
     "main.py",
     "mpvqc/__init__.py",
     "mpvqc/appearance/__init__.py",
-    "mpvqc/appearance/domain.py",
     "mpvqc/appearance/models/__init__.py",
     "mpvqc/appearance/models/accent_colors.py",
     "mpvqc/appearance/models/color_scheme_preference.py",
     "mpvqc/appearance/services/__init__.py",
     "mpvqc/appearance/services/color_scheme.py",
     "mpvqc/appearance/services/palette_catalog.py",
+    "mpvqc/appearance/services/preferences.py",
+    "mpvqc/appearance/services/schemes.py",
     "mpvqc/appearance/services/settings.py",
     "mpvqc/appearance/viewmodels/__init__.py",
     "mpvqc/appearance/viewmodels/dialog.py",
