+++ test/appearance/viewmodels/test_palette.py
@@ -8,18 +8,20 @@ from typing import NamedTuple
 import inject
 import pytest
 
-from mpvqc.appearance.domain import (
+from mpvqc.appearance.services import (
     AccentColor,
     AppearancePreference,
+    AppearanceSettingsService,
     ColorSchemePreference,
+    ColorSchemeService,
     Dark,
     FollowSystem,
     Light,
     NoPreference,
     Palette,
+    PaletteCatalogService,
     Unknown,
 )
-from mpvqc.appearance.services import AppearanceSettingsService, ColorSchemeService, PaletteCatalogService
 from mpvqc.appearance.viewmodels import (
     MpvqcPalette,
     MpvqcPaletteViewModel,
